# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me/students": {
      "post": {
        "operationId": "MyStudentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a student for the current user",
        "tags": [
          "MyStudent"
        ]
      },
      "get": {
        "operationId": "MyStudentController_findMyStudents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all students for the current user",
        "tags": [
          "MyStudent"
        ]
      }
    },
    "/api/users/me/students/{id}": {
      "get": {
        "operationId": "MyStudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "summary": "Get a student by ID for the current user",
        "tags": [
          "MyStudent"
        ]
      },
      "patch": {
        "operationId": "MyStudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "summary": "Update a student by ID for the current user",
        "tags": [
          "MyStudent"
        ]
      },
      "delete": {
        "operationId": "MyStudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Student deleted"
          },
          "404": {
            "description": "Student not found"
          }
        },
        "summary": "Delete a student by ID for the current user",
        "tags": [
          "MyStudent"
        ]
      }
    },
...
    "/api/students/{id}": {
      "get": {
        "operationId": "StudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Student"
        ]
      },
      "patch": {
        "operationId": "StudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Student"
        ]
      },
      "delete": {
        "operationId": "StudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Student"
        ]
      }
    },
...
    "/api/schedules": {
      "get": {
        "description": "Returns all schedules, optionally filtered by scheduleIds.",
        "operationId": "ScheduleController_findAll",
        "parameters": [
          {
            "name": "scheduleIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of schedule IDs to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all schedules",
        "tags": [
          "Schedule"
        ]
      },
      "post": {
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/schedules/search": {
      "get": {
        "description": "Search for schedules by pools, instructors, days of week, date, timezone, and includeReserved.",
        "operationId": "ScheduleController_search",
        "parameters": [
          {
            "name": "pools",
            "required": false,
            "in": "query",
            "description": "List of pool IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "instructors",
            "required": false,
            "in": "query",
            "description": "List of instructor IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "daysOfWeek",
            "required": false,
            "in": "query",
            "description": "List of days of the week",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Date to search for (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "required": false,
            "in": "query",
            "description": "Timezone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeReserved",
            "required": false,
            "in": "query",
            "description": "Include reserved schedules",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchScheduleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Search schedules",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/schedules/{id}": {
      "get": {
        "operationId": "ScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      },
      "patch": {
        "operationId": "ScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      },
      "delete": {
        "operationId": "ScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stansbury Swim API",
    "description": "API documentation for Stansbury Swim backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "Student name"
          },
          "birthday": {
            "type": "string",
            "description": "Birthday (ISO date string)"
          },
          "ability": {
            "type": "string",
            "description": "Ability level"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        },
        "required": [
          "name",
          "birthday",
          "ability"
        ]
      },
      "StudentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "ability": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "birthday",
          "ability",
          "notes"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Student ID"
          },
          "name": {
            "type": "string",
            "description": "Student name"
          },
          "birthday": {
            "type": "string",
            "description": "Birthday (ISO date string)"
          },
          "ability": {
            "type": "string",
            "description": "Ability level"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        },
        "required": [
          "id"
        ]
      },
      "ScheduleStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "notes"
        ]
      },
      "ScheduleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "instructor"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role"
        ]
      },
      "ScheduleRegistrationDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/ScheduleStudentDto"
          },
          "user": {
            "$ref": "#/components/schemas/ScheduleUserDto"
          }
        },
        "required": [
          "studentId",
          "userId",
          "createdAt",
          "student",
          "user"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "poolId": {
            "type": "string"
          },
          "instructorId": {
            "type": "string"
          },
          "classSize": {
            "type": "number"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleRegistrationDto"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "instructorId",
          "classSize",
          "lessonType",
          "startDateTime",
          "endDateTime",
          "registrations"
        ]
      },
...
      "SearchScheduleStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "notes"
        ]
      },
      "SearchScheduleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "instructor"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role"
        ]
      },
      "SearchScheduleRegistrationDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/SearchScheduleStudentDto"
          },
          "user": {
            "$ref": "#/components/schemas/SearchScheduleUserDto"
          }
        },
        "required": [
          "studentId",
          "userId",
          "createdAt",
          "student",
          "user"
        ]
      },
      "SearchScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "poolId": {
            "type": "string"
          },
          "instructorId": {
            "type": "string"
          },
          "classSize": {
            "type": "number"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchScheduleRegistrationDto"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "instructorId",
          "classSize",
          "lessonType",
          "startDateTime",
          "endDateTime",
          "registrations"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}